todo listoo

+++ 00-Unidades/04_while/Ejercicios_While/While_app_07.py
@@ -50,7 +50,7 @@ class App(customtkinter.CTk):
             if numero== None:
                 break
 
-            numero = float(numero)
+            numero = int(numero)
 
             
             contador +=1
